show only private and my tasks
in Wunderlist you can configure to show only tasks assigned to the current user or in private lists (no assignee). This should also be possible in WunderPebble. I can't change the WunderPebble settings but there should be some kind of setting to enable and disable this behavior and then the change in the pull request should be added